update entity UserProjectRole and Role

+++ src/main/java/com/software/service/RoleService.java
@@ -5,9 +5,9 @@ import com.software.domain.user.Role;
 import java.util.List;
 
 public interface RoleService {
-    Role getRoleById(Integer roleId);
+    Role getRoleById(Long roleId);
     Role getRoleByName(String roleName);
     List<Role> getAllRoles();
     Role creteRole(Role role);
-    void deleteRole(Integer roleId);
+    void deleteRole(Long roleId);
 }
