+++ src/main/java/com/software/service/exception/role/RoleNotFoundException.java
@@ -4,7 +4,7 @@ public class RoleNotFoundException extends RuntimeException {
     private static final String ROLE_NOT_FOUND_ID_MESSAGE = "Role with id %d not found";
     private static final String ROLE_NOT_FOUND_NAME_MESSAGE = "Role with name %s not found";
 
-    public RoleNotFoundException(Integer roleId) {
+    public RoleNotFoundException(Long roleId) {
         super(ROLE_NOT_FOUND_ID_MESSAGE.formatted(roleId));
     }
